Bug 1378586 P2 Avoid Timeout mNestingLevel rollover by just limiting the value to the values we care about. r=farre
UltraBlame original commit: adc6dc3c8c4e99dc7d50174df4a6b40b5cade227

+++ dom/base/Timeout.h
@@ -79,7 +79,9 @@ public:
   
   uint32_t mFiringId;
 
-  uint32_t mNestingLevel;
+  
+  
+  uint8_t mNestingLevel;
